feat(args): move args to arg_utils; add max_days_after_shipping argument
